fix(analytics): address review feedback
