+++ app/test/e2e/specs/flows.spec.ts
@@ -117,10 +117,10 @@ describe('Workflows create → run → inspect (real UI flow)', () => {
     await waitForTestId('flow-canvas-page', 15_000);
     await waitForTestId('flow-canvas-run', 15_000);
 
-    // All clickable analytics use the shared TrackedInteraction contract.
+    // Shared buttons expose analytics through the typed Button contract.
     // Assert the stable, content-free identifier reaches the real DOM rather
     // than leaking the generated flow id into the analytics dimension.
-    const runButton = await browser.$('[data-testid="flow-canvas-run"]');
+    const runButton = await waitForTestId('flow-canvas-run', 15_000);
     expect(await runButton.getAttribute('data-analytics-id')).toBe('flow-canvas-run');
   });
